+++ Libraries/Network/Packets/Structures/Position.hpp
@@ -16,8 +16,8 @@
  */
 struct packet_position {
     uint16_t id; // id of the entity to update position
-    uint16_t x;
-    uint16_t y;
+    float x;
+    float y;
     uint16_t index; // for 3D effect
 };
